fix(fedora-test): fix inspec puppet package name
When puppet was not installed with
official puppetlabs repos, the package name was
"puppet-agent" on fedora 29 and superior, puppet
package is installed from fedora repos and the
correct name is "puppet". The name was not set
correctly in inspect tests.

This commit change the package name test by
inspec when it run on fedora 29 and superior.