feat(docker): replace the version name rolling by latest
Only alpine, archlinux, clearlinux and gentoo
images are concerned. Previously the version
used for these os was rolling because it was
the development model used by these distros.
Behind that we used the docker tag "latest"
to retrieve this images so used "rolling" in
docker images name is not really explicit.

BREAKING CHANGE: This commit replaces in the
docker image names the suffix "rolling" by
"latest".